Cast dimensions to size_t

+++ include/data/FluidTensor.hpp
@@ -780,7 +780,7 @@ namespace fluid {
          **/
         template<typename... Dims,
         typename = enable_if_t<is_index_sequence<Dims...>()>>
-        FluidTensorView(T* p,size_t start, Dims...dims):m_desc(start,{dims...}),m_ref(p){}
+        FluidTensorView(T* p,size_t start, Dims...dims):m_desc(start,{static_cast<size_t>(dims)...}),m_ref(p){}
         
 //        /***********
 //         Construct from a whole FluidTensor
